time to lint
